Squ1rrel_CTF_2026/squ1rrel2026CTF/soulmate/starting_files/extracted/challenge/requirements.txt
setuptools<78
torch==2.1.0
torchvision==0.16.0
numpy>=1.24,<2
pyyaml
scipy
click
requests
ninja
Pillow
fastapi
uvicorn[standard]
python-multipart
transformers==4.41.2
